requirements.txt
blinker==1.8.2
click==8.1.8
colorama==0.4.6
ffmpeg-python==0.2.0
flask==3.0.3
future==1.0.0
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.5
MarkupSafe==2.1.5
pillow==10.4.0
werkzeug==3.0.6
zipp==3.20.2
